+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Change some legal vote wordings for german (!880)
 - Show no more outdated meetings in dashboard home (!859)
+- Rework protocol UX (!866)
 - Change UX for whiteboard access (!841)
 - Wording coffee break (!862)
 - Fix icons missing in chat field (!854)